+++ api/app/src/main/java/ru/osiptsoff/newspaper/api/repository/CommentRepository.java
@@ -11,13 +11,13 @@ import ru.osiptsoff.newspaper.api.model.Comment;
 import ru.osiptsoff.newspaper.api.model.News;
 
 @Repository
-public interface CommentRepository extends JpaRepository<Comment, Integer> {
+public interface CommentRepository extends JpaRepository<Comment, Long> {
     @Query("SELECT c FROM Comment c WHERE c.news.id = :id ORDER BY c.postTime DESC")
-    Page<Comment> findByNewsId(@Param("id") Integer newsId, Pageable pageable);
+    Page<Comment> findByNewsId(@Param("id") Long newsId, Pageable pageable);
     Page<Comment> findAllByNewsOrderByPostTimeDesc(News news, Pageable pageable);
 
     @Query("SELECT c.author.login FROM Comment c WHERE c.id = :id")
-    String getAuthorLogin(@Param("id") Integer commentId);
+    String getAuthorLogin(@Param("id") Long commentId);
 
     Long deleteAllByNews(News news);
 }
