+++ api/app/src/main/java/ru/osiptsoff/newspaper/api/repository/NewsContentRepository.java
@@ -14,7 +14,7 @@ import ru.osiptsoff.newspaper.api.model.embeddable.NewsContentBlockId;
 @Repository
 public interface NewsContentRepository extends PagingAndSortingRepository<NewsContentBlock, NewsContentBlockId> {
     @Query("SELECT b FROM NewsContentBlock b WHERE b.newsContentBlockId.newsId = :id ORDER BY b.newsContentBlockId.number ASC")
-    Page<NewsContentBlock> findByNewsId(@Param("id") Integer newsId, Pageable pageable);
+    Page<NewsContentBlock> findByNewsId(@Param("id") Long newsId, Pageable pageable);
 
     Long countByNews(News news);
