+++ api/app/src/main/java/ru/osiptsoff/newspaper/api/repository/TokenRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import ru.osiptsoff.newspaper.api.model.auth.Token;
 
 @Repository
-public interface TokenRepository extends CrudRepository<Token, Integer> {
+public interface TokenRepository extends CrudRepository<Token, Long> {
     Long deleteByValue(String value);
     Optional<Token> findByValue(String value);
     Boolean existsByValue(String value);
