+++ api/app/src/main/java/ru/osiptsoff/newspaper/api/service/TagService.java
@@ -79,7 +79,7 @@ public class TagService {
         }
     }
 
-    public void deleteTag(Integer id) {
+    public void deleteTag(Long id) {
         log.info("Got request to delete tag with id = " + id);
 
         try {
